Slight Cleanup
